Fix wiered issue with Inventory.DataEntries.
Inventory.DataEntries is magically set to 17 of entries when start game
play. Even I remove all the data adding process in C++.

But If I change the DataEntries into other name, the problem goes away. Or
I can change it to protected or private

I thought there was some hidden blueprint code somewhere on the stage
causing the issue. But it does not give me any compilation error in
Editor.

I guess, it is related to a engine bug?